+++ project/index.ts
@@ -1,3 +1,4 @@
 export { Paths } from "./paths"
 export { TsGdProject } from "./project"
 export { default } from "./project"
+export * from "./errors"
